Parallelize code and prepare for pretty printing
These changes do a few things that are important. We first parallelize
the code so that recursive traversal of large files will be
significantly faster using rayon, with negligible perf hit in the small
file case. We also switch our JsonPath type to use im::Vector so that
clones will be cheap as we need those rather than &mut to work in
parallel. We also setup a channel to receive messages in parallel across
all the threads that might be spun up. What we send is the new Diff type
which we can later use to setup pretty printing of the data with say
a table or something like that.