Add testing infrastructure to stop manual checks
Up to this point we've been testing the code by looking at the output.
This is awful and inneficient and harder now that the code is
parallelized. To fix this we split the code specifically for checking
into it's own library and keep the `main.rs` file for just running the
cli tool. We then test that the results we get back are what we expect.
Further work will need to be done to cleanup the setup code, probably via
a macro, but still progress!