Add pretty printing of tables to divergent
This commit changes how we display our code. Whereas before it was a bit
unclear as to what was going on, this change presents the code as an
easy to read table that's always sorted by the path itself, meaning it
will always have the same output despite being parallelized. While this
won't be the final form of the diff output, for where the project
currently is, it's a solid intermediate step. We also add testing to
make sure we visually do not regress as well.